Updated constants and started using them.

+++ MAMF45/Assets/Scripts/Illnesses/SneezeIllness.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public abstract class SneezeIllness : Illness {
-	public abstract int GetSneezeIntervalMax ();
-	public abstract int GetSneezeIntervalMin ();
+	public abstract float GetSneezeIntervalMax ();
+	public abstract float GetSneezeIntervalMin ();
 }
